requirements.txt
aiofiles==25.1.0
aiogram==3.24.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
attrs==25.4.0
bcrypt==5.0.0
certifi==2026.1.4
cffi==2.0.0
click==8.3.1
cryptography==46.0.4
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.128.0
fastapi-users==15.0.3
frozenlist==1.8.0
greenlet==3.3.1
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
magic-filter==1.0.12
makefun==1.16.0
MarkupSafe==3.0.3
multidict==6.7.1
passlib==1.7.4
propcache==0.4.1
pwdlib==0.3.0
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
rsa==4.9.1
six==1.17.0
sqladmin==0.22.0
SQLAlchemy==2.0.46
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
Werkzeug==3.1.5
WTForms==3.1.2
yarl==1.22.0
